refactor: move `Traverser` to `crate::path`, remove unused path utils, change mods visibility
