--- src/fs/mod.rs
@@ -1,3 +1,0 @@
-pub use traverser::*;
-
-mod traverser;
