+++ src/path/mod.rs
@@ -1,5 +1,5 @@
 pub use clean::*;
-pub use utils::*;
+pub use traverser::*;
 
 mod clean;
-mod utils;
+mod traverser;
