+++ src/path/traverser.rs
